seliuk-18-07-lecture The correct project structure has been made and a new test (validLogin) has been added
