+++ seliukProj/src/main/java/pages/ActionsWithElements.java
@@ -3,14 +3,16 @@ package pages;
 import org.junit.Assert;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.PageFactory;
 
 public class ActionsWithElements {
     protected WebDriver webDriver;
 
     public ActionsWithElements(WebDriver webDriver) {
         this.webDriver = webDriver;
+        PageFactory.initElements(webDriver, this); // this - means all elements from this class will be initialized
+        // elements in @FindBy annotations
     }
-
 
     public void openPage(String url) {
         try {
@@ -49,14 +51,13 @@ public class ActionsWithElements {
         }
     }
 
+    public void checkElementDisplayed(WebElement element) {
+        Assert.assertTrue("Element is not displayed", isElementDisplayed(element));
+    }
+
     private void printErrorAndStopTest(Exception e) {
         System.out.println("Can not work with element" + e);
         Assert.fail("Can not work with element" + e);
     }
-
-
-
-
-
 
 }
