+++ seliukProj/src/main/java/pages/LoginPage.java
@@ -1,0 +1,37 @@
+package pages;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+
+public class LoginPage extends ParrentPage {
+    @FindBy(xpath = "//input[@placeholder='Username']")
+    private WebElement inputUserName;
+
+    @FindBy(xpath = "//input[@placeholder='Password']")
+    private WebElement inputPassword;
+
+    @FindBy(xpath = "//button[@class='btn btn-primary btn-sm']")
+    private WebElement buttonSignIn;
+
+
+    public LoginPage(WebDriver webDriver) {
+        super(webDriver);
+    }
+
+    public void openLoginPage() {
+        openPage(BASE_URL);
+    }
+
+    public void enterUserName(String userName) {
+        enterTextIntoInput(inputUserName, userName);
+    }
+
+    public void enterPassword(String password) {
+        enterTextIntoInput(inputPassword, password);
+    }
+
+    public void clickOnButtonSignIn() {
+        clickOnElement(buttonSignIn);
+    }
+}
