+++ seliukProj/src/main/java/pages/PageProvider.java
@@ -1,0 +1,15 @@
+package pages;
+
+import org.openqa.selenium.WebDriver;
+
+public class PageProvider {
+    WebDriver webDriver;
+
+    public PageProvider(WebDriver webDriver) {
+        this.webDriver = webDriver;
+    }
+
+    public LoginPage getLoginPage() {
+        return new LoginPage(webDriver);
+    }
+}
