+++ seliukProj/src/main/java/pages/ParrentPage.java
@@ -1,0 +1,12 @@
+package pages;
+
+import org.openqa.selenium.WebDriver;
+
+public class ParrentPage extends ActionsWithElements {
+    final String BASE_URL = "https://qa-complexapp.onrender.com";
+
+    public ParrentPage(WebDriver webDriver) {
+        super(webDriver);
+    }
+
+}
