+++ seliukProj/src/main/resources/log4j.properties
@@ -1,0 +1,19 @@
+# Log levels
+log4j.rootLogger=info, stdout, R
+
+# Console appender configuration
+log4j.appender.stdout=org.apache.log4j.ConsoleAppender
+# Pattern to output the caller's file name and line number.
+log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
+log4j.appender.stdout.layout.ConversionPattern=%d{yyyy-dd-MM HH:mm:ss}  %5p [%20c{1}:%-4L] : %m%n
+
+# Rolling File Appender
+log4j.appender.R=org.apache.log4j.RollingFileAppender
+# Path and file name to store the log file.
+log4j.appender.R.File=./log/qalight.log
+log4j.appender.R.MaxFileSize=500KB
+# Keep one backup file
+log4j.appender.R.MaxBackupIndex=1
+# Rolling File Appender layout
+log4j.appender.R.layout=org.apache.log4j.PatternLayout
+log4j.appender.R.layout.ConversionPattern=%d{yyyy-dd-MM HH:mm:ss}  %5p [%c] : %m%n
